Subclass Mail from DryRunMail and mark read/unread
I meant to break these out into separate commits, but they kind of went
hand in hand.  I started working on adding the ability to mark messages
as read and unread (i.e. 'S' flag and no 'S' flag), also as new/old
(i.e. 'new' and 'cur').  This required me to redo the flag logic.  I
modeled it directly off of the Python mailbox.MailDirMessage class.

Since I was redoing _get_flagpart, I decided just to move things around
to make testing a little bit easier.  As a result, DryRunMail is not a
subclass of Mail.  All of DryRunMail's instance methods have been
removed.  The only thing it does is to remove the actions associated
with the _methods.  Also, as part of this move, I also moved the actions
parts of the Mail class methods into the _methods.  Mail.move() is a
specific case.